Add ASGI and WSGI support with worker configuration
- Introduced new `asgi.go` and `wsgi.go` files for handling ASGI and WSGI applications respectively.
- Implemented worker support for both ASGI and WSGI, allowing users to specify the number of workers in the Caddyfile.
- Updated existing tests to include worker configurations for various Python frameworks (Django, FastAPI, Flask, etc.).
- Enhanced error handling and resource management in the new implementations.
- Improved documentation to reflect the new worker configuration options.
