+++ tests/simple_async/Caddyfile
@@ -9,6 +9,7 @@ localhost:9080 {
 	route /item/* {
 		python {
 			module_asgi "main:app"
+			workers 1
 			venv "./venv"
 		}
 	}
